Modularize Promoter: replace promote_all_links legacy function
This is the next step in the modularization of the promoter code. The
promote_all_links functions has been bypassed in the new code and
reimplement in a modularized fashion.

As the promote_all_links is one of the biggest legacy function, with
responsibilities that cover the whole promoter workflow, the resulting
single step is way bigger than may be desirable, but there's no other
way to avoid this size.

The legacy promote_all_links has been altered only to add labels to the
various sections, so their equivalent implementation in the new code
can be easily identified.

Once again, the core logic of the promoter has not been modified, only
reorganized to avail of the new config object and to follow the single
reponsibility paradigm.
The various responsibilities of the legacy function have been spread
across 4 new files and classes, that in turn import legacy functions,
but prepare for the upcoming component pipeline promotion changes.

Unfortunately, this means that unit and functional tests are still too
difficult to implement at this stage, as the legacy code was never meant
to be testable.

Change-Id: Id00eb72f82f3ed62fabb3eb4cc4b7c09e88b7316