Modularize Promoter: replace setup_logging
This small change replaces the setup logging function and reorganize the
main code to create a root class Promoter that will drive the rest of
the process,on which to setup the main logging.

The workflow will now log also in stdout if it detects that it's run in
a tty.

Change-Id: I713795977de968a5fe04035f920680730285513c