Modularize Promoter: replace fetch_hashes and fetch_jobs
This change starts digging deeper into the logic, the fetch_hashes and
fetch_jobs function taht interact with dlrn server are replaces with
version that support the new config model and design.
Much of the hashes handling that was done directly in the function has
been moved to a dedicate class DlrnHash, that implaments all possible
operations around dlrn hashes.

An analogous DlrnAggregatedHash has been created to deal with the
component pipeline aggregated hash instead. This will also make
fetch_jobs and fetch_hashes polymorphism easier to deal with.

Change-Id: Idfca30160ec94f5665a765cd97e8416f57142565