Modularize Promoter: replace fetch_hashes
This change starts digging deeper into the logic, the fetch_hashes
function that interact with dlrn server is replaced with version that
support the new config model and design.
Much of the hashes handling that was done directly in the function has
been moved to a dedicated class DlrnHash, that implements all possible
operations around dlrn hashes.

An analogous DlrnAggregatedHash has been created to deal with the
component pipeline aggregated hash instead. This will also make
polymorphism easier to deal with.

This is safe to merge as no part of the workflow uses this yet

Change-Id: Idfca30160ec94f5665a765cd97e8416f57142565