Modularize Promoter: replace tag_qcow_images
This change replaces the legacy tag_qcow_images function. The logic has
not been changed, fortunately the bash script doesn't rely on the format
of the hash to perform the promotion, so we canuse the same logic for
both single and component pipeline.

No much tests can be done, as the main part is a bash script here.

Change-Id: I2e64094e0aa05608352a4a29002b5e960888a6ea