Modularize Promoter: replace tag_containers
This change replaces the tag containers legacy function, moving it to
the registry client.
Operation that are not dependent on the hash/label are grouped in the
init method, and a branch is created to deal with the component type.
As the container-push playbook has not yet been modified to support
component pipeline, the workflow will not support component pipeline

Change-Id: I03cd831eaacbcc3470bfc4d91b04c4d1a9234033