Update patch set 1
Patch Set 1:

Playbooks are harder to test than roles. By putting more code into playbooks we make it easier to break things.

The idea of a reusable playbook is bit of contradiction with the purpose of a role.

On a role at least I know that need to run its tests. But with a playbook, I will be forced to consider all its consumers.


My take is: if you cannot use a role to encapsulate reusable functionality, maybe you do not really need that reusability in the first place.

Note: I am against this direction but not strong in my conviction, it would be ok to ignore my concern.

Patch-set: 1