Update patch set 1
Patch Set 1:

> Playbooks are harder to test than roles. By putting more code into
 > playbooks we make it easier to break things.
 > 
 > The idea of a reusable playbook is bit of contradiction with the
 > purpose of a role.
 > 
 > On a role at least I know that need to run its tests. But with a
 > playbook, I will be forced to consider all its consumers.
 > 
 > 
 > My take is: if you cannot use a role to encapsulate reusable
 > functionality, maybe you do not really need that reusability in the
 > first place.
 > 
 > Note: I am against this direction but not strong in my conviction,
 > it would be ok to ignore my concern.

i agree on your notion about playbooks harder to test than roles. but these are special playbooks.

they are the ones we wire up in our zuul job definitions as pre and post and run plays 

i think really what you're asking for is a 'run promoter molecule tests' role 

imo that is tangential to this duplication. We can do both. Let's remove this duplication as a first step. 

It could then be moved to a role as another step is there is enough momentum to do so.

Patch-set: 1