Handle errors pulling ppc64le containers when creating manifests
As described in the bug below even when a given ppc tagged
container exists in the registry it can fail to be pulled.
This reworks the manifest code so we can continue when any
given ppc container fails to pull. A manifest is still created
but containing only the x86 image.

Change-Id: Ic22a2a6c8c088dceb90c68fed29c0295bf3f149a
Closes-Bug: https://bugs.launchpad.net/tripleo/+bug/1861342