[WIP] Implement promotion for integration pipeline
Promoter is currently working with a single definition of dlrn hash.
With the coming of the component pipeline, the basic assumptions for all
dlrn hashes are not true anymore.
The component pipeline introduced the concept of aggregate hash, which
does adds an attribute to the existing hash, which is computed from a
repo file through md5sum.
Known pipelines will still promote the dlrn hahses
Component pipeline will need to promote aggregated hashes

The workflow needs to take this difference into consideration to promote
for the component pipeline.
The difference in dlrnapi is rendered somewhat transparent as the api
calls for this type of hash will be the same, but overloaded with
different attributes.
To render the difference completely transparent to the whole workflow
and not repeat e.g. a check "Is this single or component pipeline?"
everywhere in the workflow, and having to test multiple execution paths
for every single answer to this question, this patch moves all the
complexity to the DlrnHash class and its tests.
The workflow should never need to address the type of pipeline directly,
and work only with the DlrnHash object in all cases.
Stage environment setup now is able to spawn and kill dlrn server from
the stage config file, and we can pass alternative fixture files.

Change-Id: I221f4b33cf2763875fc6394902f7923108a34d2a