WIP Add ppc_manifests switch to the manifest_push tasks
This adds a switch that will enable or disable the inclusion of ppc
containers in the manifest push logic.

It also adds an extra molecule test for the container push scenario
to cover the new variable. Some of the common tasks are moved into
an included tasks file to avoid repetition.

Change-Id: Ia9847c2ea4591e5b2026c6893afef94b910c223e