Add ppc_manifests check and switch to manifest_push to exclude ppc
Adds ppc_manifests switch (default true) to enable/disable the
ppc containers in the manifest push logic. If ppc_manifests is false
then only x86 manifests are created.

An attempt is made to pull N containers (the ppc_pull_attempts var
is also added so its configurable) with ppc64le tag. If any of those
fail we flip ppc_manifests false.

It also adds an extra molecule test for the container push scenario
to cover the new variable. Tracked by tripleo-ci squad in [1]

Depends-On to fix the failing promotion-staging job

Depends-On: https://review.opendev.org/707054
[1] https://tree.taiga.io/project/tripleo-ci-board/task/1511
Change-Id: Ia9847c2ea4591e5b2026c6893afef94b910c223e