Add symlink for converge to playbook.yml for molecule delegated
Due to problems with latest molecule version (>3.05a) we have
issue with missing 'playbook.yml' (newer version renames this to
converge.yml). Lets try to deal until it is fixed using links.

You can see example of failure at [1] - see zuul report for v1
and that patch comes on top of this.

[1] https://review.rdoproject.org/r/#/c/24945/1

Change-Id: I2b958ae101d042e1924932a78053bd196194f507