Update patch set 4
Patch Set 4: Code-Review+2

It was really difficult to review this, without any description on the strategy you adopted, and since infra-tripleo is really similar to tripleo-infra existing tenant it took some time to understand that you were not replacing the original configuration. It took some time to understand why you commented some part of the configuration, and without any tests, It's hard to see anything wrong with the configuration at a glance. I understand it would really difficult to test this as it needs an openstack instance, but having automated checks on the final status of the provisioning  that run even just locally it would be better. I'll just have to trust your limited manual tests in production.

Patch-set: 4
Reviewer: Gerrit User 63 <63@270e2033-b340-4cff-9539-693957ebf0e7>
Label: Code-Review=+2