Promoter: fix pipeline type default
pipeline_type default parameter was incorrectly tied to the
target_registries_push config option. As RHEL8 has the option turned
off, promotion on it was failing to pass the correct pipeline type to
dlrn operations.
This patches ties the default to the correct parameter (which at the
moment is not specified in any configuration)

Change-Id: I179a38aec8de9c9da7ddf42ac4b13779955bd5e7