[WIP] Promoter: use api call for batch aggregate promotion
https://softwarefactory-project.io/r/#/c/17748 and
https://softwarefactory-project.io/r/#/c/17749

introduced a new api call to promote a set of commitdistro hashes
associated to an aggregate hash in a single call.
This patch uses the new call as the only way to promote aggregate hash.
As the default call is single now, the implementation has been slightly
simplified merging two methods and removing one.

Change-Id: I48da64dc8584736bb1517a9f40c72d89683015f9