Promoter: fix aggregate candidate location
Dlrn promotion currently calculates the URL to get delorean repo for and
aggregate hash in the worst possible way.

we currently use http://trunk-repo/tripleo-ci-testing/delorean.repo.

The staiging environment is conveniently constructed so the only
tripleo-ci-testing we have is also the last that promoted. In real life
scenario tripleo-ci-testing may have changed a lot by the time we are
trying to promote.

This patch fixes the URL construction so it directly refers to the
correct hash:

http://trunk-repo/tripleo-ci-testing/ab/bc/abbc/delorean.repo

Unit test were testing the wrong assumption that the first URL was
correct and have nove been fixed.
Also functional tests were limiting the noise of commits, and eerything
was passing because of the unreal reproduction of scenario.
Staging environment now injects some more noise and the aggregate that
needs to promote is easily identified as the last.
Functional tests now have been improved to identify issues of missing
promotion with more string checking

Change-Id: Ibfbf154552d9c58c1a444592f41ae4f08e19fbea