Promoter: fix aggregate candidate location
Dlrn promotion is calculating the URL for the candidate hash using only
the candidate label, which is the wrong location, as the delorean.repo
in a candidate label may have changed in the meantime.
This patch fixes the location using the dlrn_hash commit_dir location,
starting from the candidate label.
Fixed also the tests that were coherent with the wrong assumption.
Functional tests may have helped find this bug if they ran more than 1
promotion per round. A change in the staging environment is needed to
add these kind of functional tests.

Change-Id: Ibfbf154552d9c58c1a444592f41ae4f08e19fbea