Promoter: Adapt provision/staging for centos 8
This change converts the promoter provision role and the
staging jobs to work on CentOS8.

The provision role is currently not working, and test for it was
disabled a while ago.

The changes break staging jobs, so these are being converted to
CentOS8 as well in the same review.

The change is big, but is mostly moving around existing code and
reorganizing tests.

- A specific delegated job was added for the provision role.
- It drops backwards compatibility completely for CentOS7.
- The tasks in main were modularized to be able to be called from
  different other roles/playbooks that need specific functionality.
- The testing has been changed to use tox, now local testing and
  upstream testing can be run with the same command.

https://tree.taiga.io/project/tripleo-ci-board/us/1569

Co-Authored-By: Rafael Folco <rfolco@redhat.com>

Depends-On: https://review.rdoproject.org/r/26030
Change-Id: I8b4e3ef494d7cabfeeca3409f93e452354c8cd63