Update train promotion criteria with tempest jobs
With this change we are updating train promotion criteria
replacing centos7 Standalone-full-tempest to two new jobs:-

* periodic-tripleo-ci-centos-7-standalone-full-tempest-api-train
* periodic-tripleo-ci-centos-7-standalone-full-tempest-scenario-train

Standalone-full-tempest was disabled from promotion criteria here[0]

We added new jobs in openstack-periodic-latest-released pipeline[1]

We made this change because Standalone-full-tempest sometimes goes out
in timeout state due to race conditions and concurrency issue. So we
have splitted it in two jobs to run faster and get better results.
Job defination is here[2]

We successfully tested run of above jobs for stable/train in
testproject here[3]

[0] https://review.rdoproject.org/r/#/c/26211/
[1] https://review.rdoproject.org/r/#/c/26228/
[2] https://review.rdoproject.org/r/#/c/26216/
[3] https://review.rdoproject.org/r/#/c/26217/

Depends-On: https://review.rdoproject.org/r/#/c/26228/
Change-Id: I8823a4912d1575461a3164b1c0e45ad4e0f2b865