[WIP] Promoter: Set broken jobs as non-voting
We're in the process of migrating jobs to centos8. This change
should give us the indication which ones are currently broken.

Jobs that are working will remain as voting and will be moving
one by one to centos8.

Broken jobs will be marked as non-voting and we'll be switching
to voting again when we bootstrap and get them fixed on centos8.

https://tree.taiga.io/project/tripleo-ci-board/us/1569

Change-Id: Ic69d28f452d43c39aee5c0ddedb6f9cef40e6a0f