Update patch set 1
Patch Set 1:

Anyway, I don't agree with how these tests are considered expendable, and disable the voting is one of the preferred way to move on, even when a solution was mentioned. It's like only 2 people have are resposible to maintain them, and it they are not around get rid of the jobs. I see one of the tests is failing on a change for Centos8/master.ini. That file cannot be waived. It's used by the staging environment to run test, if a change on that file is not tested properly it may cause problems.

Patch-set: 1