ppc64le: Remove Non-Power Registries for Containers Build Job
This patch adds a sed command to remove the existing container
registry line in registries.conf and specifies docker.io as the
only registry to pull from instead, as it actually servers
containers for power by default. Without doing this, the cico node
will instead attempt to use registry.centos.org, which is x86 only,
resulting in all containers failing to build.

Change-Id: Ie96b15606485268bf5ba2738ee38be7b47cec612