Promoter: make config root a variable
To prepare running the promoter on multiple environments, we need first
to be able to specify multiple configuration directories.

This patch adds a new option for the promoter to pass the config root.
It also makes moves the list of releases to promote, from being
hardcoded in the start script, to be an export in a config file inside
the default config root

Change-Id: I818c7a4ebc5f32af933acf8fcb25046f66139edd