Permanently remove all tests.
When a test breaks, people that clearly don't realize what the benefits
of a test are, wait for someone else to fix it, and if it's not fixed
immediately, they enjoy disabling it so their precious urgent
no-tests-needed-this-works-trust-me patch can be merged. It takes time
and energy to maintain the tests running, and only an otherwise
unwarranted self merged patch to disable them, so it's a game that is
not worth playing anymore.
It's really easy to say "the test is not relevant to my code" when you
don't have any tests for your code at all.

This patch removes all the tests permanently, so we can all merge
untested and untestable patches quickly and without supervision. The old
scheme of testing and breaking things in production worked so well,
let's not change it.

Change-Id: I9691b584b077f2d9f4301b4cd590db89f3cbd951