Update patch set 1
Patch Set 1: Code-Review+1

This looks like it's working so I say we go this route!

Folco mentioned that they had to login and push as a post step in the past as podman didn't maintain a session. This seems to work as well!

Patch-set: 1
Reviewer: Gerrit User 514 <514@270e2033-b340-4cff-9539-693957ebf0e7>
Label: Code-Review=+1