Promoter new configuration engine: Add Core
This is part of a series of patch to add the new configuration engine to
the promoter.

Patch 1: Add the new configuration core

This patch adds only the basic piece of the new configuration, the
core class that implements the basic strategy

The engine is not wired in the flow in this patch. So nothing in the
production flow will use the new engine until the engine is fully
tested.

Change-Id: I527f069ccf1143ccbef1659ff43ab7fbe9844f24