Promoter new configuration engine: Add PromoterConfig
This is part of a series of patch to add the new configuration engine to
the promoter.

Patch 2: Add the PromoterConfig class

This patch adds the class that will contain the promoter configuration.
It's a subclass of ConfigCore and specifies the layers, constructors and
filters used by promoter configuration

The engine is not wired in the flow in this patch. So nothing in the
production flow will use the new engine until the engine is fully
tested.

Task: https://tree.taiga.io/project/tripleo-ci-board/task/1743
Change-Id: Ie34a85bb1ff91ef622bdd81829a6dbfc769872aa