Promoter new configuration engine: Add PromoterConfigFactory
This is part of a series of patch to add the new configuration engine to
the promoter.

Patch 3: Add the PromoterConfigFactory

This patch adds the factory class that will be responsible to create
PromoterConfig instances. Will be the primary interface with the rest of
the promoter code, will load config layers, and validate the created
config

The engine is not wired in the flow in this patch. So nothing in the
production flow will use the new engine until the engine is full tested.

Task: https://tree.taiga.io/project/tripleo-ci-board/task/1743
Change-Id: I7684ad5f706ce62a30aec19e282b133bac6ec816