Promoter new configuration engine: stage environment
This is part of a series of patch to add the new configuration
engine to
the promoter.

Patch 4: Move staging environment to use the new engine

This patch refactor the staging environment configuration to
use the new configuration engine and classes.
Other adjustments are made to the rest of the staging code

The engine is not wired in the flow in this patch. So nothing in the
production flow will use the new engine until the engine is fully
tested.

Change-Id: I4e0957d3bb135e5d11471592bb0a6c4651135634
Task: https://tree.taiga.io/project/tripleo-ci-board/task/1743
Signed-off-by: Amol Kahat <amolkahat@gmail.com>