Promoter new configuration engine: functional tests
This is part of a series of patch to add the new configuration engine to the
promoter.

Patch 5: Add option to use new config in fixture

This patch adds option for the fixture and parameters in functional
tests to use the new configuration engine.
Two parameters are added for each test, so for the period we need to
support two configuration method, the time required for all the tests
will double

The engine is not wired in the flow in this patch. So nothing in the•
production flow will use the new engine until the engine is fully
tested.

co-authored-by: Marios Andreou <marios@redhat.com>
co-authored-by: Amol Kahat <amolkahat@gmail.com>

Task: https://tree.taiga.io/project/tripleo-ci-board/task/1743
Change-Id: Ic34dbd9d11f90d42c94f0c47fd583bd1b987189f