Promoter: download and use the containers build exclude list
The container build process has introduced a list of containers to exclude from
build that changes the expectation on containers to promote.
The list of containers is now different from the official list in tripleo-common.

This change download the file that contains the list of exclusions, and uses it
to adapt the list of containers to promote

Change-Id: Idd525b2e0e446ada60ebd5ad0bd58a4960a03e75