Promoter: fetch containers from tripleo_containers.yaml and .j2 file
With introduction to new container workflow,
container-images/tripleo_containers.yaml is the source of truth for
providing container images.
Earlier overcloud_containers.yaml.j2 was the source of getting
containers images which is still used for older release.

In order to handle both the container source, we are changing
the internal logic.
Earlier, it uses regex to get the container names.

Since the data structure of both the container source files
are different so we are removing regex dependency and moving
to jinja.

Where one file is download and read by url, then Jinja will
convert any of the file passed to Template and render it.
The rendered stream will convert it into yaml structure where
it will filter the result based on different keys present in
the source file and returns the list of containers.

An user can choose the different source file by setting
containers_list_path in criteria config file.

The complete problem statement is defined here:
https://hackmd.io/3rAE3QawRXewcC8BIEvSAg

Change-Id: If0f6c349b915b530cf554fc91505cff55c4c0d04
Signed-off-by: Chandan Kumar (raukadah) <chkumar@redhat.com>