Promoter: fetch containers from tripleo_containers.yaml also
With introduction to new container workflow,
container-images/tripleo_containers.yaml is the source of truth for
providing container images.
Earlier overcloud_containers.yaml.j2 was the source of getting
containers images which is still used for older release.

In order to make it working, we need to add both the flow
without breaking the existing one.

It modifies the repo_client.get_containers_list method to retrive
the list based on .j2 as well as .yaml file.

It also update the CentOS 8 promoter config.

The complete design is defined here:
https://hackmd.io/3rAE3QawRXewcC8BIEvSAg

Change-Id: If0f6c349b915b530cf554fc91505cff55c4c0d04
Signed-off-by: Chandan Kumar (raukadah) <chkumar@redhat.com>