Added base_container_images var to include base containers
https://github.com/rdo-infra/ci-config/blame/master/ci-scripts/container-push/roles/containers-promote/tasks/main.yaml#L48
adds the base containers from which all other containers are
built using kolla workflow and get pushed to docker.io.

As per the new container workflow, all containers are built from
RHUI so that we do not need to add base container images
for master.

Due to that, master push containers to docker.io was failing with
no openstack-base containers found. Moving it to
var fixes the issue.

Adding the vars in train, ussuri, master .ini file

Change-Id: I7166c60af95a51665cc89f29aa33b2e58cf9b932
Signed-off-by: Chandan Kumar (raukadah) <chkumar@redhat.com>