Added implicit_container_images var to add base containers
In master release, we have container build tree which has
clear relationship between root and branches in build tree
in new container workflow. Base container used here is RHUI
and it gets pulled from RH registry and that donot needs to
be built or pushed to docker.io.

In kolla drived workflow, we still need to
append implicit images {base, openstack-base} to containers
list so that they gets pushed to registry otherwise promoter
will fail to push those containers.

Moving the implicit images to a var make it easier to manage
through configuration.

Change-Id: I7166c60af95a51665cc89f29aa33b2e58cf9b932
Signed-off-by: Chandan Kumar (raukadah) <chkumar@redhat.com>