Run mol-promoter as unprivileged user
Switch the promoter role ro run as a unprivileged user.
This avoids environment nuances around users and permissions
on delegated e non-delegated tasks. Now the promoter scenario
is non-delegated and runs in a remote instance on molecule
local testing. On zuul env instance is 127.0.1.1, localhost.
This means no longer running docker on docker like in legacy
jobs. Credentials are generated locally on controller, and
copied to the instance, under promoter's home dir.
dlrn service has been moved to a template that executes
script on the ci-config path (no more hard coded file).

Job is NOT green yet, it fails at the end of verify step.
This is going to be fixed in follow-up patches.

Change-Id: I889bef75e64bae05a4c2255421e9c2ced69cf403