Promoter: Get container list based on image_source
tripleo-common/container-images/{overcloud_containers.yaml.j2,
tripleo_containers.yaml} are the sources of containers.

overcloud_containers.yaml.j2 is based on kolla driven workflow &
having image_source is kolla, ceph and prom.
and tripleo_containers.yaml is based on new container workflow
& having image_source is tripleo, ceph, graphana and prom.

We only builds container whose image source is kolla or tripleo
and push it to rdo registry.

So while pushing to container, using image_source to generate the
container list gives a exclusive list of containers which
are actually built and also helps it from not to maintain
exclude list of containers not coming from other sources.

Change-Id: I1506969502568ac277a99cdc82409782694e9f43
Signed-off-by: Chandan Kumar (raukadah) <chkumar@redhat.com>