Remove release variable from get_hash converge
Mentioning release when running get_hash scenario is wrong because
it will fail on centos7 which does not accept master builds.

It is better to keep it default, so the role can decide which
release to pick for current platform.

This change also fixes the conditions for component jobs versus
legacy ones. Adding a distro condition to the molecule scenarios
until we split them up.

Change-Id: I72f4c4e8f04dc07b49e951ac054d7cf09df29827