Promoter: use {container,overcloud}_containers.yaml as container source
Earlier  overcloud_containers.yaml.j2 file is used for containers
list but makes the link between THT services and their container
images. So we are not able to retrive correct containers and carry
a lot of hacks for parsing j2 file.

However, overcloud,tripleo_containers.yaml provides the list of
images in a release and where to pull them by default.

This patch drops .j2 file and switch to .yaml file.
It also updates the ussuri and train criterias in order
to point to correct container file path.

It is not going to break anything as overcloud_containers.yaml
as is present in all stable/releases, can be pulled with
tripleo-common commit hash used in promotion pipeline.

Queens release does not have image_source in yaml but
that case is also handled.

It also introduces build_method to highlight which container
build method is used. Default is kolla for
overcloud_containers.yaml file and tripleo for
tripleo_containers.yaml.

Tracked here:
https://tree.taiga.io/project/tripleo-ci-board/task/1895

Change-Id: I1e9296584f1a096071a99d3a5f77f64b9dee679d
Signed-off-by: Ronelle Landy <rlandy@redhat.com>
Co-authored-by: Chandan Kumar (raukadah) <chkumar@redhat.com>