Create new role _ensure_credentials
The credentials are created directly on target path in the
remote instance so we don't need to copy them over.

Create this private role to share code across all molecule
scenarios and zuul jobs that need to run promoter provisioning
role in a test/staging environment. And we get code de-dup as
a bonus.

Change-Id: I6506d70fc94ad7c483fa6dc3c14c9a2be225dcd2