Update patch set 23
Patch Set 23:

> (1 comment)
 > 
 > Since queens, rocky and stein will still use the old image build
 > workflow, you cannot replace the method, but you need to create
 > another method in parallel, and maintain both, so you need a
 > variable to identify which build method to use. We already
 > discussed this a bit in my last email to you.

I have added the build_method. For older workflow it is based on kolla and newer one is based on tripleo.
In order to use overcloud_containers.yaml. Only for queens, 
https://opendev.org/openstack/tripleo-common/src/branch/stable/queens/container-images/overcloud_containers.yaml Image_source as kolla is not defined which is handled and for other release
schema is same. We have added the code and unit tests to handle all cases. It will work fine.

Patch-set: 23
Label: Code-Review=0