Re enable standalone ci centos gate job
Support for standaone is added as part of lp#1889219,
so let's enable the job.

Add 'check-ci-centos' comment to be used to rerun just
ci centos job on gerrit changes.

Also remove ansible-role-collect-logs usage from job
as two projects can't be run with same gerrit review.
It can be enabled on collect logs as followup.

Related-Bug: lp#1889219
Change-Id: Ida81c07384a6d432aa4e1e60820b2244e4507fdf