ppc64le: Update python requirements
The container build job is currently only buillding a fraction of
containers. We suspect that this could be because of the lower
version of ansible we are using compared to upstream, or the
missing psutil package [0]. This patch moves the ansible version
up to match upstream and installs psutil as well.

[0] https://logserver.rdoproject.org/ci.centos.org/tripleo-upstream-containers-build-master-ppc64le/2815/logs/logs/build.log

Change-Id: Ibb2e8033500f829201b63eca405eb1d0bb39272f