Adding tool to copy containers
This is a tool to copy containers between one registry to another.
This will be used to copy current-tripleo containers from docker
to quay, but it can be used to other purposes as well.
The code is written in golang, and uses the podman api, so is
faster than use python rest-api, and can be used in different
podman versions.

Change-Id: I30bdf46b98e8a68999aee068d05f9855e2aea5f4