Added --config-root flag
It will help to specify the environment type so that one
can identify whether it is a staging , rdo or downstream or
any other environments.

Otherwise we have to change the DEFAULT_CONFIG_ROOT in the
production code while deploying it.

It can also be toggled via setting promoter_config_root var
in promoter role.

Change-Id: Id60bb16badf330e248ccd93e11722434d8401938
Signed-off-by: Chandan Kumar (raukadah) <chkumar@redhat.com>