Bump cryptography version
https://github.com/pyca/cryptography/issues/5771 broke the
whole ci, it needs to be updated. Since it is only
available in test-requirements.txt, by running tox -e deps, it
does not update the cryptography version, so changing the version
manually to 3.3.2 as per the suggestion given in
https://github.com/Azure/azure-cli/issues/16858
Also added cryptography module in
ci-script/dlrnapi_promoter/requirements.txt

Make sure pip version >=19.1.1 is installed.

Change-Id: I9ca231a9aff8434613e7d7a71a95767401420615
Signed-off-by: Chandan Kumar (raukadah) <chkumar@redhat.com>
Co-authored-by: Bhagyashri Shewale <bshewale@redhat.com>
Co-authored-by: Amol Kahat <amolkahat@gmail.com>