ppc64le: Install greenlet from source for image and container jobs
Currently we install greenlet from source for the container images
jobs, but we should also do this for the disk images job.
Also, we should insert greenlet at the top of each list rather
than at the end as the name of the list is less likely to change
than its contents.

Change-Id: I08f30ce28520a7473c089ba477429bc520b55f4b