Copy required files to zuul log directory
Due to massive change in the promoter code, the staging log
files are dumped in /tmp.
During post log collection, there is no files in /home/zuul
so there is no logs in CI.

During staging promoter run, it will copy the artifact to
promoter home directory and then it will synced to log
server.

Change-Id: I95b57a75a45a520504e3679eb475c04333173490
Signed-off-by: Chandan Kumar (raukadah) <chkumar@redhat.com>