Fixed container-push multiple registry playbook
This patch fixes container-push multiple registry
playbook as target_registeries expects a list
containing all the resources related to registry,
This patch re-arranges back with proper list and
refactor the generate_list.sh script to avoid
the error during promotion with other registires.

Note: This change is tested in the promoter server
and it's working fine.

Closes-Bug: #1923221
Change-Id: Id184246f877b47d0e8cb1624ec8da92f0fb68912