To fix cloud and region showing up as null in rrcockpit
Issue - Cloud and Region appears as null in all tables in Rrcockpit.

Reason:
hosts.keys() generates a dictionary (example in case of multinode - {"primary", "secondary"} )
and not a list.
So [0] doesnt work and it goes into Exception and passes (without writing any logs).
So cloud, region and provider values are not registered in the db.

Questions:
1. Preserving the original format I have changed hosts.keys() to list(hosts.keys()), but it can also be rewritten as host = hosts["primary"]. Would that be better, or list(...) is?
2. Should the Exception print something on console instead of pass?

Change-Id: I986bb2d7fabf9c879f562ab509f1137daf5f12ff