Update patch set 1
Patch Set 1:

sorry I didn't get to this in time yesterday, but
why are we doing that? The plan is no wallaby jobs (and in fact removal of most of the jobs from all branches) for *upstream* but we keep them in the components and integration lines?

let's discuss later on our calls

Patch-set: 1
Reviewer: Gerrit User 192 <192@270e2033-b340-4cff-9539-693957ebf0e7>