Fix the manifest push containers task
Fixed the manifest push containers task.
As the "pulled_ppc.results" is the list of dictionary
(the structure is given below) so we can not apply
the map function on it.
This patch refactored the condition and removed the map() use.

The "pulled_ppc" value is as below:
pulled_ppc: {"results": [{"failed": "false"}, {"failed": "false"}]}

This change is tested on production server and it's working:
https://hub.docker.com/r/tripleomaster/openstack-base/tags?page=1&ordering=last_updated

Change-Id: I9c213c08203648a3088f8df484b048704e6b55ba