Improvement in ppc containers manifest push design
container-promote role will push ppc manifest if all
the container's ppc manifests are present in the
source registry, else it will disable it.
This patch will check the container manifest present
or not using shell script and will generate the
files. Which help us to identify missing ppc
manifests of the containers.

Change-Id: I5513d74d5d01bfa5329964295baef30436b8201d
Co-authored-by: Amol Kahat <amolkahat@gmail.com>