[wip]Ruck/rover tool
Adding two script:-

* ruck_rover.py
This script gives list of failing jobs which need to pass to promote
an integration hash/ component line.

* promotion_status.py
This script gives tracking of when was a particular branch promoted
and if new content is available

The data ruck_rover.py creates will be picked by telegraf to populate
influxdb to be consumed by grafana.
Adding jobs_blocking_promotion.conf to run every 2 hours

Change-Id: Iacbad17a0952f93aca98d9986a6f4e31e6e9a68e