[WIP] Use extended hash for downsteram promotion
Currently we are facing below issue with upstream dlrn promotion:
ERROR    promoter Exception while promoting hashes to API endpoint
(404) NOT FOUND: commit_hash+distro_hash+extended_hash combination
not found for
4ac3d795199c0e654249030838dc5ab764a75745_da427891168e302bd6ed805343b3cf20adaecdf5_None

To overcome this issue, we come up with two solutions as given below:
1. Remove the extended hash option from modules [1]
2. Use extended hash only for downstream promotion. [2] i.e this change

Also made changes in the unit and functional tests
which will test the dlrn promotion with commit and distro hash
and not the extended hash. We will add the separate unit and
functional tests to test the extended hash changes in the
follow patches.

[1]: https://review.rdoproject.org/r/c/rdo-infra/ci-config/+/34125
[2]: https://review.rdoproject.org/r/c/rdo-infra/ci-config/+/34124

Change-Id: Iacd04dba40944eb3fa982453ee0d0f74ab436231