Fix skiplist go script
Due the fact skiplist is a binary and the system might not be the same
as the binary, it is important to build the skiplist in the target
machine.
This patch remove the binary code and create a task to install golang
and build the skiplist binary directly on the target machine, avoiding
errors with different glib versions.

Change-Id: I4a836b42a54f244f6d0e7fc3dd247ab0ed1df1fd