Update patch set 40
Patch Set 40:

Here in the log [1] the manifest of ppc containers are not getting inspected because we don't have all the manifest for ppc containers on the stage server but that will be getting tested in the patch [2].
Note: I have tested this change in our production environment so we can merge this change to unblock our other task. +1 from my side. 

[1]: https://logserver.rdoproject.org/15/33815/40/check/mol-container_push_manifests/3038496/tox/reports.html
[2]: https://review.rdoproject.org/r/c/rdo-infra/ci-config/+/34334

Patch-set: 40
Reviewer: Gerrit User 670 <670@270e2033-b340-4cff-9539-693957ebf0e7>